Moved the UI for location needed to components. Created Tabbar, AddMeal, LandingScreen modules and moved some other things to their relevant modules
