+++ iosHackthonApp/Modules/Features/AddMeal/README.md
@@ -1,0 +1,3 @@
+# AddMeal
+
+A description of this package.
