+++ iosHackthonApp/Modules/Features/LandingScreen/README.md
@@ -1,0 +1,3 @@
+# LandingScreen
+
+A description of this package.
