+++ iosHackthonApp/Modules/Features/TabBar/README.md
@@ -1,0 +1,3 @@
+# TabBar
+
+A description of this package.
