+++ iosHackthonApp/iosHackthonApp.xcworkspace/contents.xcworkspacedata
@@ -8,6 +8,15 @@
       location = "container:"
       name = "Features">
       <FileRef
+         location = "group:Modules/Features/AddMeal">
+      </FileRef>
+      <FileRef
+         location = "group:Modules/Features/TabBar">
+      </FileRef>
+      <FileRef
+         location = "group:Modules/Features/LandingScreen">
+      </FileRef>
+      <FileRef
          location = "group:Modules/Features/MealList">
       </FileRef>
    </Group>
